Enable go-to-definition for cross-module imports
Populate the WorkspaceIndex with all discovered modules from the
package when a file is opened. This allows go-to-definition to work
for imported symbols without requiring all files to be open in the
editor.

For example, clicking on `gcd` in `use pkg.math_utils.{gcd}` will
now jump to the definition in math_utils.ab.